Начат проект EventManager итерация 1
